requirements.txt
lxml>=3.8
sphinxcontrib-napoleon
python-dateutil
pytz
termcolor
graphviz
typing
pytest
ipy
coverage
edxml-schema==3.0.0.dev3
edxml-test-corpus==3.0.0.dev4
edxml-bricks-computing>=3.0.0.dev0
edxml-bricks-computing-networking>=3.0.0.dev0
edxml-bricks-generic>=3.0.0.dev0
